Add better alias logic.
Updated alias resolution logic to ensure that the instantiated object
is persisted internally to both the original alias and the classpath
that it resolves to.

Additionally, that alias is removed from the list of aliases
as it is no longer needed and will result in there being less
aliases to check when the next one is requested.